`FilterScreen` ui components names refactoring.
