+++ app/src/main/java/ua/blackwind/limbushelper/ui/screens/filter_screen/drawer_sheet/FilterSinnersElements.kt
@@ -24,7 +24,7 @@ import ua.blackwind.limbushelper.ui.util.getSinnerIconById
 private const val NUMBER_OF_GRID_ROWS = 3
 
 @Composable
-fun FilterSinnersBlock(
+fun FilterSinnersUiContainer(
     state: FilterSinnersBlockState,
     onSinnerCheckedChange: (FilterSinnerModel) -> Unit
 ) {
